requirements.txt
Django
mysqlclient
python-dotenv
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-rest-passwordreset
channels
daphne
channels_redis